[mgsim-refactor] Fix reporting of 1-byte sampling variables.
git-svn-id: https://svn.svp-home.org/sw/svp/mg/mgsim/branches/refactor@4473 e97e5017-a994-416e-809d-76780e9f78db

+++ sim/sampling.cpp
@@ -71,7 +71,7 @@ void ListSampleVariables_onevar(std::ostream& os, const std::string& name, const
         switch(vinfo.type) {
         case SV_INTEGER:
             switch(vinfo.width) {
-            case 1: os << std::dec << *(uint8_t*)p; break;
+            case 1: os << std::dec << (unsigned)*(uint8_t*)p; break;
             case 2: os << std::dec << *(uint16_t*)p; break;
             case 4: os << std::dec << *(uint32_t*)p; break;
             case 8: os << std::dec << *(uint64_t*)p; break;
@@ -124,7 +124,7 @@ bool ReadSampleVariables(std::ostream& os, const std::string& pat)
         switch(vinfo.type) {
         case SV_INTEGER:
             switch(vinfo.width) {
-            case 1: os << std::dec << *(uint8_t*)p; break;
+            case 1: os << std::dec << (unsigned)*(uint8_t*)p; break;
             case 2: os << std::dec << *(uint16_t*)p; break;
             case 4: os << std::dec << *(uint32_t*)p; break;
             case 8: os << std::dec << *(uint64_t*)p; break;
